Releasing version 3.9.0
