+++ requirements.txt
@@ -14,7 +14,7 @@ Jinja2==3.0.3
 jmespath==0.10.0
 ndg-httpsclient==0.4.2
 mock==2.0.0
-oci==2.66.0
+oci==2.67.0
 packaging==20.2
 pluggy==0.13.0
 py==1.10.0
